+++ decidim-surveys/app/commands/decidim/surveys/admin/update_survey.rb
@@ -39,7 +39,8 @@ module Decidim
               position: form_question.position,
               mandatory: form_question.mandatory,
               question_type: form_question.question_type,
-              answer_options: form_question.answer_options_to_persist.map { |answer| { "body" => answer.body } }
+              answer_options: form_question.answer_options_to_persist.map { |answer| { "body" => answer.body } },
+              max_choices: form_question.max_choices
             }
 
             if form_question.id.present?