+++ decidim-surveys/spec/shared/edit_survey_examples.rb
@@ -199,17 +199,31 @@ shared_examples "edit surveys" do
 
     it "preserves answer options form across submission failures" do
       click_button "Add question"
-      select "Single option", from: "Type"
+      select "Multiple option", from: "Type"
 
       within ".survey-question-answer-option:first-of-type" do
         fill_in find_nested_form_field_locator("body_en"), with: "Something"
       end
+
+      click_button "Add answer option"
+
+      within ".survey-question-answer-option:last-of-type" do
+        fill_in find_nested_form_field_locator("body_en"), with: "Else"
+      end
+
+      select "3", from: "Maximum number of choices"
 
       click_button "Save"
 
       within ".survey-question-answer-option:first-of-type" do
         expect(page).to have_nested_field("body_en", with: "Something")
       end
+
+      within ".survey-question-answer-option:last-of-type" do
+        fill_in find_nested_form_field_locator("body_en"), with: "Else"
+      end
+
+      expect(page).to have_select("Maximum number of choices", selected: "3")
     end
 
     it "allows switching translated field tabs after form failures" do
@@ -226,6 +240,50 @@ shared_examples "edit surveys" do
         expect(page).to have_nested_field("body_en", with: "Bye")
         expect(page).to have_no_selector(nested_form_field_selector("body_ca"))
         expect(page).to have_no_content("Adeu")
+      end
+    end
+
+    context "when adding a multiple option question" do
+      before do
+        visit_component_admin
+
+        within "form.edit_survey" do
+          click_button "Add question"
+
+          within ".survey-question" do
+            fill_in find_nested_form_field_locator("body_en"), with: "This is the first question"
+          end
+
+          expect(page).to have_no_content "Add answer option"
+          expect(page).to have_no_select("Maximum number of choices")
+        end
+      end
+
+      it "updates the max choices selector according to the configured options" do
+        select "Multiple option", from: "Type"
+        expect(page).to have_select("Maximum number of choices", options: %w(Any 2))
+
+        click_button "Add answer option"
+        expect(page).to have_select("Maximum number of choices", options: %w(Any 2 3))
+
+        click_button "Add answer option"
+        expect(page).to have_select("Maximum number of choices", options: %w(Any 2 3 4))
+
+        within(".survey-question-answer-option:last-of-type") { click_button "Remove" }
+        expect(page).to have_select("Maximum number of choices", options: %w(Any 2 3))
+
+        within(".survey-question-answer-option:last-of-type") { click_button "Remove" }
+        expect(page).to have_select("Maximum number of choices", options: %w(Any 2))
+
+        click_button "Add question"
+
+        within(".survey-question:last-of-type") do
+          select "Multiple option", from: "Type"
+          expect(page).to have_select("Maximum number of choices", options: %w(Any 2))
+
+          select "Single option", from: "Type"
+          expect(page).to have_no_select("Maximum number of choices")
+        end
       end
     end
 
@@ -264,6 +322,7 @@ shared_examples "edit surveys" do
             fill_in "survey_questions_#{survey_question.id}_body_en", with: ""
             check "Mandatory"
             select "Multiple option", from: "Type"
+            select "2", from: "Maximum number of choices"
           end
 
           click_button "Save"
@@ -275,6 +334,7 @@ shared_examples "edit surveys" do
         expect(page).to have_selector("input[value='']")
         expect(page).to have_no_selector("input[value='This is the first question']")
         expect(page).to have_selector("input#survey_questions_#{survey_question.id}_mandatory[checked]")
+        expect(page).to have_select("Maximum number of choices", selected: "2")
         expect(page).to have_selector("select#survey_questions_#{survey_question.id}_question_type option[value='multiple_option'][selected]")
       end
 
@@ -426,7 +486,7 @@ shared_examples "edit surveys" do
 
       context "when moving a question up" do
         before do
-          within "#survey_question_#{survey_question_2.id}-field" do
+          within ".survey-question:last-of-type" do
             click_button "Up"
           end
 
@@ -436,7 +496,7 @@ shared_examples "edit surveys" do
 
       context "when moving a question down" do
         before do
-          within "#survey_question_#{survey_question_1.id}-field" do
+          within ".survey-question:first-of-type" do
             click_button "Down"
           end
         end
@@ -451,7 +511,7 @@ shared_examples "edit surveys" do
         expect(page.find(".survey-question:nth-child(2)")).to look_like_intermediate_question
         expect(page.find(".survey-question:nth-child(3)")).to look_like_last_question
 
-        within "#survey_question_#{survey_question_1.id}-field" do
+        within ".survey-question:first-of-type" do
           click_button "Remove"
         end